requirements.txt
Jinja2>=3.1.1,<4
pandas>=1.4.2,<1.5
python-slugify>=6.1.1,<7
requests>=2.27.1,<3
